requirements.txt
django==2.0.2
requests==2.18.4
python-weixin==0.3.0
djangorestframework==3.7.7
jsonfield==2.0.2
